requirements.txt
openai
flask
injective-py
hypercorn
colorama
python-dotenv
quart
pyyaml
quart-cors